move get_token_ids_logprobs_batch_optimized to logprob_processor
